requirements.txt
asgiref==3.10.0
Django==5.2.7
django-allauth==65.12.0
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
pillow==11.3.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
